main/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
aiofiles==23.2.1 
docker==6.1.3
python-dotenv==1.0.0
google-generativeai==0.3.1
pefile==2023.2.7 
oletools
pdfminer